Remove the filters from metricsPlugin.ts and place them in filters/*
